FIXED: performance issue
Root cause of the issues with performance was the during the crossover. I had implemented a loop that attempts to find a different value to crossover with if the same value is picked. Since crossing over the same value with itself is rather redundent (the result will be two vectors exactly the same as before). However, since the selection process is roulette wheel and therefore members that preform better are more likely to be selected, as the population improves the selection may select the same chromosome numerous times in a row. The loop to attempt to get another chromosome has been removed. Alternative, if the two chromosomes selected for crossover are the same they are clonned over instead of crossing to the same end.