FIXED: another source of poor performance.
As part of the design of the roulette wheel, it will iterate through the values in the map until it finds the range in which the random number is contained. However, in order to place values into the range they were inserted in the 'init' method however the map was never cleared. Therefore the map size continued to grow for each call to the 'breed' method. This would therefore increase the delay of look up if for values inserted into the map (more values in the map more to iterate through). Added a clear at the beginning of the 'init' method which significantly improves the performance.