FIXED: Access issue #2
The root of this was because a thread was pushing values into a vector that may end up not corresponding to the values the thread was working on. With no wait the thread would then proceed to access the vector at the pre-defined index which may be unset (since other workers are not done).