FIXED: population convergence to bad value.
The random number generator for the roulette wheel was set by with an upper and lower bound. The lower bound (0) and the upper bound (the sum of all the fitness values for the master population). However, the upper bound variable (Roulette::right) was not reset after each round. Therefore the random number generator would produce numbers within the range that do not map to a fitness function. The default action of no match found is to return 0. Therefore after a few generations the the population would converge to the value stored at the first index.